Fix interface parameter names
