+++ server/routes.ts
@@ -1,7 +1,8 @@
 import type { Express, Request, Response } from "express";
 import { createServer, type Server } from "http";
 import { storage } from "./storage";
-import { insertGameSchema, insertPlayerSchema } from "@shared/schema";
+import { insertGameSchema, insertPlayerSchema, insertUserSchema } from "@shared/schema";
+import bcrypt from "bcrypt";
 import { ZodError } from "zod";
 
 // This handler sets up API routes and serves the React app
@@ -9,6 +10,76 @@ const setupRoutes = (app: Express, server: Server) => {
   // Health check
   app.get('/api/health', (req, res) => {
     res.json({ status: 'ok' });
+  });
+
+  // Authentication routes
+  app.post('/api/signup', async (req, res) => {
+    try {
+      const data = insertUserSchema.parse(req.body);
+      if (!/^[a-zA-Z0-9]{3,20}$/.test(data.username)) {
+        return res.status(400).json({ error: 'Invalid username' });
+      }
+
+      const existing = await storage.getUserByUsername(data.username);
+      if (existing) {
+        return res.status(400).json({ error: 'Username already taken' });
+      }
+
+      const hashed = await bcrypt.hash(data.password, 10);
+      const user = await storage.createUser({
+        username: data.username,
+        password: hashed,
+      });
+      res.status(201).json({ id: user.id, username: user.username });
+    } catch (err) {
+      if (err instanceof ZodError) {
+        return res.status(400).json({ error: err.errors });
+      }
+      console.error('Signup error:', err);
+      res.status(500).json({ error: 'Failed to create user' });
+    }
+  });
+
+  app.post('/api/login', async (req, res) => {
+    const { username, password } = req.body as {
+      username?: string;
+      password?: string;
+    };
+    if (!username || !password) {
+      return res.status(400).json({ error: 'Missing credentials' });
+    }
+
+    try {
+      const user = await storage.getUserByUsername(username);
+      if (!user) {
+        return res.status(401).json({ error: 'Invalid username or password' });
+      }
+      const valid = await bcrypt.compare(password, user.password);
+      if (!valid) {
+        return res.status(401).json({ error: 'Invalid username or password' });
+      }
+      res.json({ id: user.id, username: user.username });
+    } catch (err) {
+      console.error('Login error:', err);
+      res.status(500).json({ error: 'Failed to login' });
+    }
+  });
+
+  app.get('/api/profile/:id', async (req, res) => {
+    const id = Number(req.params.id);
+    if (Number.isNaN(id)) {
+      return res.status(400).json({ error: 'Invalid id' });
+    }
+    try {
+      const user = await storage.getUser(id);
+      if (!user) {
+        return res.status(404).json({ error: 'User not found' });
+      }
+      res.json({ id: user.id, username: user.username });
+    } catch (err) {
+      console.error('Profile fetch error:', err);
+      res.status(500).json({ error: 'Failed to fetch profile' });
+    }
   });
 
   // Game routes
