+++ server/storage.ts
@@ -16,13 +16,13 @@ export interface IStorage {
   // Game methods
   getGame(id: number): Promise<Game | undefined>;
   getGames(): Promise<Game[]>;
-  getGamesBySport(sport: string): Promise<Game[]>;
+  getGamesBySport(skillLevel: string): Promise<Game[]>;
   createGame(game: InsertGame): Promise<Game>;
   
   // Player methods
   getPlayer(id: number): Promise<Player | undefined>;
   getPlayers(): Promise<Player[]>;
-  getPlayersBySport(sport: string): Promise<Player[]>;
+  getPlayersBySport(skillLevel: string): Promise<Player[]>;
   createPlayer(player: InsertPlayer): Promise<Player>;
 }
 
@@ -62,11 +62,11 @@ export class DatabaseStorage implements IStorage {
     return db.select().from(games);
   }
   
-  async getGamesBySport(sport: string): Promise<Game[]> {
+  async getGamesBySport(skillLevel: string): Promise<Game[]> {
     return db
       .select()
       .from(games)
-      .where(eq(games.sport, sport));
+      .where(eq(games.skillLevel, skillLevel));
   }
   
   async createGame(game: InsertGame): Promise<Game> {
@@ -90,11 +90,11 @@ export class DatabaseStorage implements IStorage {
     return db.select().from(players);
   }
   
-  async getPlayersBySport(sport: string): Promise<Player[]> {
+  async getPlayersBySport(skillLevel: string): Promise<Player[]> {
     return db
       .select()
       .from(players)
-      .where(eq(players.sport, sport));
+      .where(eq(players.skillLevel, skillLevel));
   }
   
   async createPlayer(player: InsertPlayer): Promise<Player> {
